+++ lib/pars_validator.dart
@@ -1,6 +1,6 @@
 library pars_validator;
 
-export 'src/extensions.dart';
+export 'src/extensions/extensions.dart';
 export 'src/models/models.dart';
 export 'src/validators/validators.dart';
 export 'src/widgets/widgets.dart';